Feature/10 validation

See merge request p7037/technokratos_practice!15

+++ technokratos-practice-airbnb/airbnb-impl/src/main/java/com/technokratos/service/impl/ApartmentPhotoServiceImpl.java
@@ -3,12 +3,12 @@ package com.technokratos.service.impl;
 import com.technokratos.exception.*;
 import com.technokratos.model.ApartmentEntity;
 import com.technokratos.model.ApartmentPhotoEntity;
-import com.technokratos.model.UserEntity;
 import com.technokratos.repository.ApartmentPhotoRepository;
 import com.technokratos.repository.ApartmentRepository;
 import com.technokratos.repository.PhotoRepository;
 import com.technokratos.service.ApartmentPhotoService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -30,45 +30,54 @@ public class ApartmentPhotoServiceImpl implements ApartmentPhotoService {
     private final ApartmentRepository apartmentRepository;
 
     @Override
-    public String upload(MultipartFile photo, UUID apartmentId) {
-        validatePhoto(photo);
-        try {
-            ApartmentEntity apartmentEntity = apartmentRepository.findById(apartmentId)
-                    .orElseThrow(ApartmentNotFoundException::new);
-            String id = photoRepository.savePhoto(photo);
-            ApartmentPhotoEntity apartmentPhotoEntity = ApartmentPhotoEntity.builder()
-                    .apartment(apartmentEntity)
-                    .photoId(id)
-                    .build();
-            apartmentPhotoRepository.save(apartmentPhotoEntity);
-            return id;
-        } catch (IOException e) {
-            throw new PhotoUploadException();
-        }
+    public String upload(MultipartFile photo, UUID apartmentId, UserDetails userPrincipal) {
+        ApartmentEntity apartmentEntity = apartmentRepository.findById(apartmentId)
+                .orElseThrow(ApartmentNotFoundException::new);
+        if (apartmentEntity.getOwner().getEmail().equals(userPrincipal.getUsername())) {
+            validatePhoto(photo);
+            try {
+
+                String id = photoRepository.savePhoto(photo);
+                ApartmentPhotoEntity apartmentPhotoEntity = ApartmentPhotoEntity.builder()
+                        .apartment(apartmentEntity)
+                        .photoId(id)
+                        .build();
+                apartmentPhotoRepository.save(apartmentPhotoEntity);
+                return id;
+            } catch (IOException e) {
+                throw new PhotoUploadException();
+            }
+        } else throw new AccessDeniedException();
     }
 
     @Override
-    public void delete(String photoId, UUID apartmentId) {
-        try {
-            ApartmentPhotoEntity apartmentPhotoEntity = apartmentPhotoRepository.findByPhotoId(photoId)
-                    .orElseThrow(PhotoNotFoundException::new);
-            photoRepository.findById(photoId).orElseThrow(PhotoNotFoundException::new);
-            if (apartmentPhotoEntity.getApartment().getId().equals(apartmentId)) {
-                apartmentPhotoRepository.delete(apartmentPhotoEntity);
-                photoRepository.delete(photoId);
-            } else throw new ApartmentPhotoNotFoundException();
-        } catch (IOException e) {
-            throw new PhotoNotFoundException();
-        }
+    public void delete(String photoId, UUID apartmentId, UserDetails userPrincipal) {
+        ApartmentPhotoEntity apartmentPhotoEntity = apartmentPhotoRepository.findByPhotoId(photoId)
+                .orElseThrow(PhotoNotFoundException::new);
+        if (apartmentPhotoEntity.getApartment().getOwner().getEmail().equals(userPrincipal.getUsername())) {
+            try {
+                photoRepository.findById(photoId).orElseThrow(PhotoNotFoundException::new);
+                if (apartmentPhotoEntity.getApartment().getId().equals(apartmentId)) {
+                    apartmentPhotoRepository.delete(apartmentPhotoEntity);
+                    photoRepository.delete(photoId);
+                } else throw new ApartmentPhotoNotFoundException();
+            } catch (IOException e) {
+                throw new PhotoNotFoundException();
+            }
+        } else throw new AccessDeniedException();
     }
 
     @Override
-    public void deleteAll(UUID apartmentId) {
-        apartmentPhotoRepository.findByApartment_Id(apartmentId)
-                .forEach(entity -> {
-                    photoRepository.delete(entity.getPhotoId());
-                    apartmentPhotoRepository.delete(entity);
-                });
+    public void deleteAll(UUID apartmentId, UserDetails userPrincipal) {
+        if (apartmentRepository.findById(apartmentId).
+                orElseThrow(ApartmentNotFoundException::new)
+                .getOwner().getEmail().equals(userPrincipal.getUsername())) {
+            apartmentPhotoRepository.findByApartment_Id(apartmentId)
+                    .forEach(entity -> {
+                        photoRepository.delete(entity.getPhotoId());
+                        apartmentPhotoRepository.delete(entity);
+                    });
+        } else throw new AccessDeniedException();
     }
 
     @Override
